Move `find_value_by_key()` to utils, and also move tests
